# swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"CollectionResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"collectionId": {
						"type": "string"
					}
				},
				"required": [
					"message",
					"collectionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AssetBaseObject": {
				"properties": {
					"name": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"type": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CreateCollectionRequest": {
				"properties": {
					"displayName": {
						"type": "string",
						"nullable": true
					},
					"image": {
						"allOf": [
							{
								"$ref": "#/components/schemas/AssetBaseObject"
							}
						],
						"nullable": true
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"shortUrl": {
						"type": "string"
					}
				},
				"required": [
					"shortUrl"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCollectionRequest": {
				"properties": {
					"collectionId": {
						"type": "string"
					},
					"image": {
						"allOf": [
							{
								"$ref": "#/components/schemas/AssetBaseObject"
							}
						],
						"nullable": true
					},
					"displayName": {
						"type": "string",
						"nullable": true
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"shortUrl": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"collectionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CollectionRequest": {
				"properties": {
					"collectionId": {
						"type": "string"
					}
				},
				"required": [
					"collectionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateProductResponse": {
				"properties": {
					"status": {
						"type": "string"
					},
					"transactionId": {
						"type": "string"
					},
					"productId": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"productId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AssetObject": {
				"properties": {
					"name": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"type": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"url",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateProductBody": {
				"properties": {
					"asset": {
						"$ref": "#/components/schemas/AssetObject"
					},
					"previewImage": {
						"allOf": [
							{
								"$ref": "#/components/schemas/AssetBaseObject"
							}
						],
						"nullable": true
					},
					"price": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"currency": {
						"type": "string",
						"nullable": true
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"royalties": {
						"type": "number",
						"format": "double"
					},
					"freeMinting": {
						"type": "boolean"
					},
					"draft": {
						"type": "boolean"
					}
				},
				"required": [
					"asset",
					"title",
					"royalties",
					"freeMinting",
					"draft"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProductResponse": {
				"properties": {
					"status": {
						"type": "string"
					},
					"transactionId": {
						"type": "string"
					},
					"productId": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"productId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProductBody": {
				"properties": {
					"asset": {
						"allOf": [
							{
								"$ref": "#/components/schemas/AssetObject"
							}
						],
						"nullable": true
					},
					"previewImage": {
						"allOf": [
							{
								"$ref": "#/components/schemas/AssetBaseObject"
							}
						],
						"nullable": true
					},
					"price": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"currency": {
						"type": "string",
						"nullable": true
					},
					"title": {
						"type": "string",
						"nullable": true
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"royalties": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"freeMinting": {
						"type": "boolean",
						"nullable": true
					},
					"draft": {
						"type": "boolean",
						"nullable": true
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteProductResponse": {
				"properties": {
					"status": {
						"type": "string"
					},
					"productId": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"productId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RestoreProductResponse": {
				"properties": {
					"status": {
						"type": "string"
					},
					"productId": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"productId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReactionResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReactionRequest": {
				"properties": {
					"productId": {
						"type": "string"
					}
				},
				"required": [
					"productId"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "framely-creator-service",
		"version": "1.0.0",
		"description": "Creator Service",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/collections": {
			"post": {
				"operationId": "CreateCollection",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CollectionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Collection"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCollectionRequest"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateCollection",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CollectionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Collection"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCollectionRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCollection",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CollectionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Collection"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CollectionRequest"
							}
						}
					}
				}
			}
		},
		"/collections/restore": {
			"patch": {
				"operationId": "RestoreCollection",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CollectionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Collection"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CollectionRequest"
							}
						}
					}
				}
			}
		},
		"/": {
			"get": {
				"operationId": "Index",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/products": {
			"post": {
				"operationId": "CreateProduct",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateProductResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProductBody"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateProduct",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpdateProductResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProductBody"
							}
						}
					}
				}
			}
		},
		"/products/{id}": {
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteProductResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/restore-product/{id}": {
			"patch": {
				"operationId": "RestoreProduct",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RestoreProductResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/reactions": {
			"patch": {
				"operationId": "UpdateReaction",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpdateReactionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Reaction"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReactionRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api/creator"
		}
	]
}